Use Michael's initialisation of derivative recurrence instead of Boyd's
Boyd's form neglects a[N] spectral coefficient of f, as Boyd sets the
highest two values of spectral coefficients to zero. This can leave a
rapidly oscillating component in the simulation domain which does not
advect away because it does not contribute to the derivative of f!
Michael initialises the N-1 component of f' using the a[N] component of
f, and so avoids this issue.